requirements.txt
fastapi
uvicorn
groq
python-dotenv
sqlite-utils
pydantic
ipykernel
python-multipart
streamlit
requests
